% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eerjmaster.R
\docType{data}
\name{bromelia}
\alias{bromelia}
\title{Alternative substrates for growth of bromeliads.}
\format{A data frame with 120 observations and 5 variables, where

    \describe{

    \item{\code{nleaves}}{Number of leaves per plot.}
    \item{\code{nplots}}{Number of plots.}
    \item{\code{days}}{Days after planting.}
    \item{\code{treat}}{Components to alternative substrates. The
    substrates were Xaxim, Pinus bark, Eucalyptus bark, Coxim, and
    Coir fiber, mixed with peat and perlite.}
    \item{\code{block}}{Indicates the block.}
    }}
\usage{
data(bromelia)
}
\description{
Xaxim is used in the past for containers to grow
    bromeliads and orchids whose was commercially prohibited since
    2001. Experiments have been conducted to find alternative
    container material. This dataset was produced from an experiment
    set up in a randomized complete block design with five
    alternative components to substrate pots (with turfa and perlita)
    for bromeliads. Plots with 8 plants, initially, were followed
    during 435 days after planting and the total number of leaves per
    plot was recorded at five dates.
}
\examples{

data(bromelia)
str(bromelia)
}
\keyword{datasets}
