% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eerjmaster.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{Artifitial Defoliation in Cotton Phenology}
\format{A data frame with 125 observations and 3 variables, where

    \describe{

    \item{\code{bolls}}{Number of cotton bolls.}
    \item{\code{defol}}{Levels of the artificial defoliation (percent
    in leaf area removed with scissors).}
    \item{\code{phenol}}{Phenological growth stages of the cotton
    plants.}
    }}
\usage{
data(cotton)
}
\description{
Several sources may lead to defoliation in the cotton
    crop, e.g. pests, diseases of foliage and mechanical
    injuries. This dataset comes from a greenshouse factorial
    experiment performed to evaluate the effect of artifical
    defoliation in the total number of cotton bolls at different
    growth stages. The experimental unit was a pot with 2 cotton
    plants. The response variable is the number of bolls produced at
    the end of the crop cycle.
}
\examples{

data(cotton)
str(cotton)
}
\keyword{datasets}
