% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eerjmaster.R
\docType{package}
\name{eerjmaster}
\alias{eerjmaster}
\alias{eerjmaster-package}
\title{Codes and data for master's thesis of Eduardo Jr}
\description{
Code implementation and real data used to the master's
    thesis od Eduardo Jr that proposes flexible models for count
    data.
}
